Current state

+++ internal/api/global_concurrency_limits.go
@@ -17,32 +17,32 @@ type GlobalConcurrencyLimitsClient interface {
 // GlobalConcurrencyLimit is a representation of a global concurrency limit.
 type GlobalConcurrencyLimit struct {
 	BaseModel
-	Active             bool   `json:"active"`
-	Name               string `json:"name"`
-	Limit              int64  `json:"limit"`
-	ActiveSlots        int64  `json:"active_slots"`
-	DeniedSlots        int64  `json:"denied_slots"`
-	SlotDecayPerSecond int64  `json:"slot_decay_per_second"`
+	Active             bool    `json:"active"`
+	Name               string  `json:"name"`
+	Limit              int64   `json:"limit"`
+	ActiveSlots        int64   `json:"active_slots"`
+	DeniedSlots        int64   `json:"denied_slots"`
+	SlotDecayPerSecond float64 `json:"slot_decay_per_second"`
 }
 
 // GlobalConcurrencyLimitCreate is a subset of GlobalConcurrencyLimit used when creating global concurrency limits.
 type GlobalConcurrencyLimitCreate struct {
-	Active             bool   `json:"active"`
-	Name               string `json:"name"`
-	Limit              int64  `json:"limit"`
-	ActiveSlots        int64  `json:"active_slots"`
-	DeniedSlots        int64  `json:"denied_slots"`
-	SlotDecayPerSecond int64  `json:"slot_decay_per_second"`
+	Active             bool    `json:"active"`
+	Name               string  `json:"name"`
+	Limit              int64   `json:"limit"`
+	ActiveSlots        int64   `json:"active_slots"`
+	DeniedSlots        int64   `json:"denied_slots"`
+	SlotDecayPerSecond float64 `json:"slot_decay_per_second"`
 }
 
 // GlobalConcurrencyLimitUpdate is a subset of GlobalConcurrencyLimit used when updating global concurrency limits.
 type GlobalConcurrencyLimitUpdate struct {
-	Active             bool   `json:"active"`
-	Name               string `json:"name"`
-	Limit              int64  `json:"limit"`
-	ActiveSlots        int64  `json:"active_slots"`
-	DeniedSlots        int64  `json:"denied_slots"`
-	SlotDecayPerSecond int64  `json:"slot_decay_per_second"`
+	Active             bool    `json:"active"`
+	Name               string  `json:"name"`
+	Limit              int64   `json:"limit"`
+	ActiveSlots        int64   `json:"active_slots"`
+	DeniedSlots        int64   `json:"denied_slots"`
+	SlotDecayPerSecond float64 `json:"slot_decay_per_second"`
 }
 
 // GlobalConcurrencyLimitFilter is a filter for global concurrency limits.
